SFR 450176 Slow media cut-through.
Don't bother starting a parser (which we try to stop immediately after)
when all we want is the names of the contained tracks, which doesn't
require us to start it.  Not having to stop the parser (and the underlying
source object) saves up to 400ms on slow machines (depending on thread
synchronization). Also fix some speeling and typso. Also include something
in a sync block that should be synched, rather than having it outside the
block.

(cherry picked from commit c7fdf02794e5552f98605b2f388743e27a01f5d1)